[SRC] Removing task_log_merge_stderr option (impossible to add options to [logging] section)
