+++ tests/system/nomad/config/unit_tests.cfg.template
@@ -742,7 +742,6 @@ fab_logging_level = WARNING
 # Variable: AIRFLOW__LOGGING__LOGGING_CONFIG_CLASS
 #
 logging_config_class = airflow.providers.nomad.nomad_log.NOMAD_LOG_CONFIG
-task_log_merge_with_stderr= false
 
 # Flag to enable/disable Colored logs in Console
 # Colour the logs when the controlling terminal is a TTY.
