+++ tests/unit/nomad/config/unit_tests.cfg
@@ -30,8 +30,6 @@ task_log_reader = nomad_log_handler
 #
 logging_config_class = airflow.providers.nomad.nomad_log.NOMAD_LOG_CONFIG
 
-task_log_merge_with_stderr = false
-
 [nomad]
 # Generic Airflow executor parallelism (should be higher than 0)
 #
